+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/crypto/store/db/model/DeviceInfoEntity.kt
@@ -47,3 +47,8 @@ internal open class DeviceInfoEntity(
 
     companion object
 }
+
+internal fun DeviceInfoEntity.deleteOnCascade() {
+    trustLevelEntity?.deleteFromRealm()
+    deleteFromRealm()
+}